telemetry: use at most one open db per client
Previously, each client db could be opened multiple times, with one
shared one for writing during the lifetime of the client and multiple
other ones for reading (over the sse endpoints).

We also have been seeing occasional mysterious hangs where it becomes
impossible to acquire a write lock on the DB, with `SQLITE_BUSY` errors
returned and the engine hanging indefinitely trying to write.

There's a few potential interrelated explanations:
1. While working on this PR I found out that `clientLogs.OnEmit`
   sometimes gets called after the session has shutdown and the DB was
   closed. This seems to be possible when the main client does a sudden
   cancel while there's still some containers running; those containers
   try to flush after the DB was already closed. I'm not totally sure
   what was happening before this change in that scenario, but may
   easily have entered "undefined behavior" territory since db Close may
   have been called in midst of write attempts.
2. `SQLITE_BUSY` can be triggered for a number of reasons, including
   attempts to use the DB while in the midst of shutdown, during a
   recovery attempt, etc.
   https://www.sqlite.org/wal.html#sometimes_queries_return_sqlite_busy_in_wal_mode

The change here wraps the client db with ref counting and
synchronization such that there is at most one db conn open at a time
and used by both readers+writers. It's closed when no-one has an open
ref to it and re-opened if accessed after that.

To help performance, a ref is consistently held open for the lifetime of
the client (to prevent frequent close/opens).

+++ engine/clientdb/dbs.go
@@ -1,6 +1,7 @@
 package clientdb
 
 import (
+	"context"
 	"database/sql"
 	_ "embed"
 	"errors"
@@ -9,6 +10,7 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/dagger/dagger/engine/slog"
@@ -17,6 +19,9 @@ import (
 
 type DBs struct {
 	Root string
+
+	open map[string]*DB
+	mu   sync.RWMutex
 }
 
 // CollectGarbageAfter is the time after which a database is considered
@@ -24,54 +29,120 @@ type DBs struct {
 const CollectGarbageAfter = time.Hour
 
 func NewDBs(root string) *DBs {
-	return &DBs{Root: root}
+	return &DBs{
+		Root: root,
+		open: make(map[string]*DB),
+	}
 }
 
 //go:embed schema.sql
 var Schema string
 
-// Create creates a new database for the given clientID and runs the schema
-// migration. This operation must be idempotent.
-func (dbs *DBs) Create(clientID string) (*sql.DB, error) {
-	db, err := dbs.Open(clientID)
-	if err != nil {
-		return nil, err
+// Open open a database for the given clientID if not open already and
+// runs the schema migration if needed.
+func (dbs *DBs) Open(ctx context.Context, clientID string) (_ *DB, rerr error) {
+	lg := slog.Default().With("clientID", clientID)
+
+	dbs.mu.Lock()
+	db, ok := dbs.open[clientID]
+	if !ok {
+		db = &DB{
+			dbs:      dbs,
+			clientID: clientID,
+		}
+		dbs.open[clientID] = db
 	}
-	if _, err := db.Exec(Schema); err != nil {
-		return nil, fmt.Errorf("migrate: %w", err)
+	dbs.mu.Unlock()
+
+	db.mu.Lock()
+	defer db.mu.Unlock()
+	defer func() {
+		if rerr != nil {
+			rerr = errors.Join(rerr, dbs.close(db, lg))
+		}
+	}()
+	db.refCount++
+
+	lg = lg.With("aquiredRefCount", db.refCount)
+
+	if db.inner == nil {
+		lg.ExtraDebug("opening client DB", "clientID", clientID)
+
+		dbPath := db.dbs.path(db.clientID)
+		if err := os.MkdirAll(filepath.Dir(dbPath), 0700); err != nil {
+			return nil, fmt.Errorf("mkdir %s: %w", filepath.Dir(dbPath), err)
+		}
+		connURL := &url.URL{
+			Scheme: "file",
+			Host:   "",
+			Path:   dbPath,
+			RawQuery: url.Values{
+				"_pragma": []string{
+					"foreign_keys=ON",    // we don't use em yet, but makes sense anyway
+					"journal_mode=WAL",   // readers don't block writers and vice versa
+					"synchronous=OFF",    // we don't care about durability and don't want to be surprised by syncs
+					"busy_timeout=10000", // wait up to 10s when there are concurrent writers
+				},
+				"_txlock": []string{"immediate"}, // use BEGIN IMMEDIATE for transactions
+			}.Encode(),
+		}
+		sqlDB, err := sql.Open("sqlite", connURL.String())
+		if err != nil {
+			return nil, fmt.Errorf("open %s: %w", connURL, err)
+		}
+		if err := sqlDB.Ping(); err != nil {
+			return nil, fmt.Errorf("ping %s: %w", connURL, err)
+		}
+
+		db.inner = sqlDB
+		if _, err := db.inner.Exec(Schema); err != nil {
+			return nil, fmt.Errorf("migrate: %w", err)
+		}
+
+		db.Queries, err = Prepare(ctx, db.inner)
+		if err != nil {
+			return nil, fmt.Errorf("prepare queries: %w", err)
+		}
+	} else {
+		lg.ExtraDebug("reusing open client DB", "clientID", clientID)
 	}
+
 	return db, nil
 }
 
-// Open opens the database for the given client, for reading.
-func (dbs *DBs) Open(clientID string) (*sql.DB, error) {
-	dbPath := dbs.path(clientID)
-	if err := os.MkdirAll(filepath.Dir(dbPath), 0700); err != nil {
-		return nil, fmt.Errorf("mkdir %s: %w", filepath.Dir(dbPath), err)
+// assumes db.mu is held and dbs.mu is not held
+func (dbs *DBs) close(db *DB, lg *slog.Logger) (rerr error) {
+	db.refCount--
+
+	lg = lg.With("releasedRefCount", db.refCount)
+
+	if db.refCount > 0 {
+		lg.ExtraDebug("not closing client DB; still in use")
+		return nil
 	}
-	connURL := &url.URL{
-		Scheme: "file",
-		Host:   "",
-		Path:   dbPath,
-		RawQuery: url.Values{
-			"_pragma": []string{
-				"foreign_keys=ON",    // we don't use em yet, but makes sense anyway
-				"journal_mode=WAL",   // readers don't block writers and vice versa
-				"synchronous=OFF",    // we don't care about durability and don't want to be surprised by syncs
-				"busy_timeout=20000", // wait up to 20s when there are concurrent writers
-			},
-			"_txlock": []string{"immediate"}, // use BEGIN IMMEDIATE for transactions
-		}.Encode(),
+
+	lg.ExtraDebug("closing client DB; no more references")
+
+	db.refCount = 0
+
+	if db.Queries != nil {
+		if cerr := db.Queries.Close(); cerr != nil {
+			rerr = errors.Join(rerr, fmt.Errorf("error closing queries: %w", cerr))
+		}
+		db.Queries = nil
 	}
-	db, err := sql.Open("sqlite", connURL.String())
-	if err != nil {
-		return nil, fmt.Errorf("open %s: %w", connURL, err)
+	if db.inner != nil {
+		if cerr := db.inner.Close(); cerr != nil {
+			rerr = errors.Join(rerr, fmt.Errorf("error closing db: %w", cerr))
+		}
+		db.inner = nil
 	}
-	if err := db.Ping(); err != nil {
-		db.Close()
-		return nil, fmt.Errorf("ping %s: %w", connURL, err)
-	}
-	return db, nil
+
+	dbs.mu.Lock()
+	defer dbs.mu.Unlock()
+	delete(dbs.open, db.clientID)
+
+	return rerr
 }
 
 // GC removes databases that are older than CollectGarbageAfter based on mtime.
@@ -103,6 +174,14 @@ func (dbs *DBs) GC(keep map[string]bool) error {
 			// DB is still fresh; keep
 			continue
 		}
+		dbs.mu.RLock()
+		_, openBySomeone := dbs.open[clientID]
+		dbs.mu.RUnlock()
+		if openBySomeone {
+			// DB is still open by someone; keep it but log this since this is a weird case, possibly indicative of a leak
+			slog.Warn("skipping garbage collection of client DB that is still open", "clientID", clientID)
+			continue
+		}
 		if err := os.RemoveAll(filepath.Join(dbs.Root, ent.Name())); err != nil {
 			errs = errors.Join(errs, fmt.Errorf("remove %s: %w", ent.Name(), err))
 		}
@@ -116,4 +195,26 @@ func (dbs *DBs) GC(keep map[string]bool) error {
 
 func (dbs *DBs) path(clientID string) string {
 	return filepath.Join(dbs.Root, clientID+".db")
+}
+
+type DB struct {
+	dbs      *DBs
+	clientID string
+
+	inner *sql.DB
+	*Queries
+
+	refCount int
+	mu       sync.Mutex
+}
+
+func (db *DB) Begin() (*sql.Tx, error) {
+	return db.inner.Begin()
+}
+
+func (db *DB) Close() (rerr error) {
+	lg := slog.Default().With("clientID", db.clientID)
+	db.mu.Lock()
+	defer db.mu.Unlock()
+	return db.dbs.close(db, lg)
 }
